* VB-2335 Create active in active get Session Template end point (#434)

* Vb 2335 create active in active get session template end point fix (#437)

* VB-2335 Create active in active get Session Template end point

* VB-2335 Create active in active get Session Template end point, updated some code (tidy up) and document

* VB-2335 Create active in active get Session Template end point, updated defect found by Tom

* VB-2335 Create active in active get Session Template end point, updated defect found by Tom

* VB-2335 Create active in active get Session Template end point, updated defect found by Tom

* Merge main to develop (#438)

* VB-2237 - remove Monday session template. (#428)

* Fix to resolve issues with multiple events being created (#431)

* VB-2310 Update SQS/SNS queue updates, do not release until tested in preprod env (#429)

* VB-2332 Added additional information to logs (#432)

* VB-2332 Improve logging when SNS fails... (#433)

* VB-2334 - improve utility that flag visits to send customEvents to ApplicationInsights and add reports. (#435)

---------


* VB-2237 - update dev timings to flag visits task every 1 hour. (#440)

* Vb 2341 move admin endpoints into correct code base/package (#441)

* VB-2341 Move admin endpoints into correct code base / package

* VB-2341 Move admin endpoints into correct code base / package

* Add create update delete catigory for sessions (#443)

* Added Create/Update and Delete end points for category

* Added Create/Update and Delete end points for category,
Tests were failing

* Added Create/Update and Delete end points for category,
Tests were failing

* Vb 2348 build api endpoints to enable maintenance of incentives (#444)

* VB-2348 Build API Endpoints to enable maintenance of Incentive Level Groups

* VB-2348 Build API Endpoints to enable maintenance of Incentive Level Groups, correction of typo's spotted by DH

* VB-2348 Build API Endpoints to enable maintenance of Incentive Level Groups, the normal issues with the hash function

* VB-2348 Build API Endpoints to enable maintenance of Incentive Level Groups, the normal issues with the hash function

* VB-2348 Build API Endpoints to enable maintenance of Incentive Level Groups, the normal issues with the hash function

* VB-2321 DELETE API function on session templates will not allow deletion of Session Templates with visits (#446)

* Vb 2323 add basic server side validations for session template (#445)

* VB-2223 - add basic validations around create and update session template.

* VB-2223 - update swagger docs. (#448)

* Merge main into develop (#449)

---------

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/service/SessionTemplateService.kt
@@ -4,9 +4,19 @@ import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
+import uk.gov.justice.digital.hmpps.visitscheduler.controller.admin.SessionTemplateRangeType
+import uk.gov.justice.digital.hmpps.visitscheduler.controller.admin.SessionTemplateRangeType.ACTIVE_OR_FUTURE
+import uk.gov.justice.digital.hmpps.visitscheduler.controller.admin.SessionTemplateRangeType.ALL
+import uk.gov.justice.digital.hmpps.visitscheduler.controller.admin.SessionTemplateRangeType.HISTORIC
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.CreateSessionTemplateDto
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.SessionTemplateDto
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.UpdateSessionTemplateDto
+import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.category.CreateCategoryGroupDto
+import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.category.SessionCategoryGroupDto
+import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.category.UpdateCategoryGroupDto
+import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.incentive.CreateIncentiveGroupDto
+import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.incentive.SessionIncentiveLevelGroupDto
+import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.incentive.UpdateIncentiveGroupDto
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.location.CreateLocationGroupDto
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.location.SessionLocationGroupDto
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.location.UpdateLocationGroupDto
@@ -15,15 +25,16 @@ import uk.gov.justice.digital.hmpps.visitscheduler.exception.VSiPValidationExcep
 import uk.gov.justice.digital.hmpps.visitscheduler.model.VisitType
 import uk.gov.justice.digital.hmpps.visitscheduler.model.entity.session.SessionTemplate
 import uk.gov.justice.digital.hmpps.visitscheduler.model.entity.session.category.SessionCategoryGroup
+import uk.gov.justice.digital.hmpps.visitscheduler.model.entity.session.category.SessionPrisonerCategory
 import uk.gov.justice.digital.hmpps.visitscheduler.model.entity.session.incentive.SessionIncentiveLevelGroup
+import uk.gov.justice.digital.hmpps.visitscheduler.model.entity.session.incentive.SessionPrisonerIncentiveLevel
 import uk.gov.justice.digital.hmpps.visitscheduler.model.entity.session.location.PermittedSessionLocation
 import uk.gov.justice.digital.hmpps.visitscheduler.model.entity.session.location.SessionLocationGroup
 import uk.gov.justice.digital.hmpps.visitscheduler.repository.SessionCategoryGroupRepository
 import uk.gov.justice.digital.hmpps.visitscheduler.repository.SessionIncentiveLevelGroupRepository
 import uk.gov.justice.digital.hmpps.visitscheduler.repository.SessionLocationGroupRepository
 import uk.gov.justice.digital.hmpps.visitscheduler.repository.SessionTemplateRepository
-import java.time.DayOfWeek
-import java.time.LocalDate
+import uk.gov.justice.digital.hmpps.visitscheduler.repository.VisitRepository
 import java.util.function.Supplier
 
 @Service
@@ -33,6 +44,7 @@ class SessionTemplateService(
   private val sessionLocationGroupRepository: SessionLocationGroupRepository,
   private val sessionCategoryGroupRepository: SessionCategoryGroupRepository,
   private val sessionIncentiveLevelGroupRepository: SessionIncentiveLevelGroupRepository,
+  private val visitRepository: VisitRepository,
   private val prisonConfigService: PrisonConfigService,
 ) {
 
@@ -41,14 +53,12 @@ class SessionTemplateService(
   }
 
   @Transactional(readOnly = true)
-  fun getSessionTemplates(prisonCode: String, dayOfWeek: DayOfWeek?, rangeStartDate: LocalDate?, rangeEndDate: LocalDate?): List<SessionTemplateDto> {
-    val sessionTemplates = sessionTemplateRepository.findValidSessionTemplatesBy(
-      prisonCode = prisonCode,
-      rangeStartDate = rangeStartDate,
-      rangeEndDate = rangeEndDate,
-      dayOfWeek = dayOfWeek,
-    )
-
+  fun getSessionTemplates(prisonCode: String, rangeType: SessionTemplateRangeType): List<SessionTemplateDto> {
+    val sessionTemplates = when (rangeType) {
+      ALL -> sessionTemplateRepository.findSessionTemplatesByPrisonCode(prisonCode)
+      HISTORIC -> sessionTemplateRepository.findInActiveSessionTemplates(prisonCode)
+      ACTIVE_OR_FUTURE -> sessionTemplateRepository.findActiveAndFutureSessionTemplates(prisonCode)
+    }
     return sessionTemplates.sortedBy { it.validFromDate }.map { SessionTemplateDto(it) }
   }
 
@@ -58,12 +68,12 @@ class SessionTemplateService(
   }
 
   @Transactional(readOnly = true)
-  fun getSessionLocationGroupByReference(reference: String): SessionLocationGroupDto {
+  fun getSessionLocationGroup(reference: String): SessionLocationGroupDto {
     return SessionLocationGroupDto(getLocationGroupByReference(reference))
   }
 
   @Transactional(readOnly = true)
-  fun getSessionLocationGroup(prisonCode: String): List<SessionLocationGroupDto> {
+  fun getSessionLocationGroups(prisonCode: String): List<SessionLocationGroupDto> {
     return sessionLocationGroupRepository.findByPrisonCode(prisonCode).map { SessionLocationGroupDto(it) }
   }
 
@@ -75,7 +85,7 @@ class SessionTemplateService(
       name = createLocationSessionGroup.name,
     )
 
-    val sessionLocations = createLocationSessionGroup.locations.map {
+    val sessionLocations = createLocationSessionGroup.locations.toSet().map {
       PermittedSessionLocation(
         groupId = sessionLocationGroup.id,
         sessionLocationGroup = sessionLocationGroup,
@@ -97,7 +107,7 @@ class SessionTemplateService(
     sessionLocationGroup.name = updateLocationSessionGroup.name
     sessionLocationGroup.sessionLocations.clear()
 
-    val sessionLocations = updateLocationSessionGroup.locations.map {
+    val sessionLocations = updateLocationSessionGroup.locations.toSet().map {
       PermittedSessionLocation(
         groupId = sessionLocationGroup.id,
         sessionLocationGroup = sessionLocationGroup,
@@ -122,28 +132,28 @@ class SessionTemplateService(
       prisonId = prison.id,
       prison = prison,
       name = createSessionTemplateDto.name,
-      startTime = createSessionTemplateDto.startTime,
-      endTime = createSessionTemplateDto.endTime,
-      validFromDate = createSessionTemplateDto.validFromDate,
-      validToDate = createSessionTemplateDto.validToDate,
+      startTime = createSessionTemplateDto.sessionTimeSlot.startTime,
+      endTime = createSessionTemplateDto.sessionTimeSlot.endTime,
+      validFromDate = createSessionTemplateDto.sessionDateRange.validFromDate,
+      validToDate = createSessionTemplateDto.sessionDateRange.validToDate,
       visitRoom = createSessionTemplateDto.visitRoom,
-      closedCapacity = createSessionTemplateDto.closedCapacity,
-      openCapacity = createSessionTemplateDto.openCapacity,
+      closedCapacity = createSessionTemplateDto.sessionCapacity.closed,
+      openCapacity = createSessionTemplateDto.sessionCapacity.open,
       biWeekly = createSessionTemplateDto.biWeekly,
       dayOfWeek = createSessionTemplateDto.dayOfWeek,
       visitType = VisitType.SOCIAL,
     )
 
     createSessionTemplateDto.categoryGroupReferences?.let {
-      it.forEach { ref -> sessionTemplateEntity.permittedSessionCategoryGroups.add(this.getPrisonerCategoryGroupByReference(ref)) }
+      it.toSet().forEach { ref -> sessionTemplateEntity.permittedSessionCategoryGroups.add(this.getPrisonerCategoryGroupByReference(ref)) }
     }
 
     createSessionTemplateDto.locationGroupReferences?.let {
-      it.forEach { ref -> sessionTemplateEntity.permittedSessionLocationGroups.add(this.getLocationGroupByReference(ref)) }
+      it.toSet().forEach { ref -> sessionTemplateEntity.permittedSessionLocationGroups.add(this.getLocationGroupByReference(ref)) }
     }
 
     createSessionTemplateDto.incentiveLevelGroupReferences?.let {
-      it.forEach { ref -> sessionTemplateEntity.permittedSessionIncentiveLevelGroups.add(this.getPrisonerIncentiveLevelGroupByReference(ref)) }
+      it.toSet().forEach { ref -> sessionTemplateEntity.permittedSessionIncentiveLevelGroups.add(this.getIncentiveLevelGroupByReference(ref)) }
     }
 
     val sessionTemplateEntitySave = sessionTemplateRepository.saveAndFlush(sessionTemplateEntity)
@@ -159,28 +169,19 @@ class SessionTemplateService(
         sessionTemplateRepository.updateNameByReference(reference, name)
       }
 
-      startTime?.let {
-        sessionTemplateRepository.updateStartTimeByReference(reference, startTime)
+      sessionTimeSlot?.let {
+        sessionTemplateRepository.updateStartTimeByReference(reference, it.startTime)
+        sessionTemplateRepository.updateEndTimeByReference(reference, it.endTime)
       }
 
-      endTime?.let {
-        sessionTemplateRepository.updateEndTimeByReference(reference, endTime)
+      sessionDateRange?.let {
+        sessionTemplateRepository.updateValidFromDateByReference(reference, it.validFromDate)
+        sessionTemplateRepository.updateValidToDateByReference(reference, it.validToDate)
       }
 
-      validFromDate?.let {
-        sessionTemplateRepository.updateValidFromDateByReference(reference, validFromDate)
-      }
-
-      validToDate?.let {
-        sessionTemplateRepository.updateValidToDateByReference(reference, validToDate)
-      }
-
-      closedCapacity?.let {
-        sessionTemplateRepository.updateClosedCapacityByReference(reference, closedCapacity)
-      }
-
-      openCapacity?.let {
-        sessionTemplateRepository.updateOpenCapacityByReference(reference, openCapacity)
+      sessionCapacity?.let {
+        sessionTemplateRepository.updateClosedCapacityByReference(reference, it.closed)
+        sessionTemplateRepository.updateOpenCapacityByReference(reference, it.open)
       }
 
       biWeekly?.let {
@@ -192,22 +193,26 @@ class SessionTemplateService(
 
     updateSessionTemplateDto.locationGroupReferences?.let {
       updatedSessionTemplateEntity.permittedSessionLocationGroups.clear()
-      it.forEach { ref -> updatedSessionTemplateEntity.permittedSessionLocationGroups.add(this.getLocationGroupByReference(ref)) }
+      it.toSet().forEach { ref -> updatedSessionTemplateEntity.permittedSessionLocationGroups.add(this.getLocationGroupByReference(ref)) }
     }
 
     updateSessionTemplateDto.categoryGroupReferences?.let {
       updatedSessionTemplateEntity.permittedSessionCategoryGroups.clear()
-      it.forEach { ref -> updatedSessionTemplateEntity.permittedSessionCategoryGroups.add(this.getPrisonerCategoryGroupByReference(ref)) }
+      it.toSet().forEach { ref -> updatedSessionTemplateEntity.permittedSessionCategoryGroups.add(this.getPrisonerCategoryGroupByReference(ref)) }
     }
 
     updateSessionTemplateDto.incentiveLevelGroupReferences?.let {
       updatedSessionTemplateEntity.permittedSessionIncentiveLevelGroups.clear()
-      it.forEach { ref -> updatedSessionTemplateEntity.permittedSessionIncentiveLevelGroups.add(this.getPrisonerIncentiveLevelGroupByReference(ref)) }
+      it.toSet().forEach { ref -> updatedSessionTemplateEntity.permittedSessionIncentiveLevelGroups.add(this.getIncentiveLevelGroupByReference(ref)) }
     }
     return SessionTemplateDto(updatedSessionTemplateEntity)
   }
 
   fun deleteSessionTemplates(reference: String) {
+    if (visitRepository.hasVisitsForSessionTemplate(reference)) {
+      throw VSiPValidationException("Cannot delete session template $reference with existing visits!")
+    }
+
     val deleted = sessionTemplateRepository.deleteByReference(reference)
     if (deleted == 0) {
       throw ItemNotFoundException("Session template not found : $reference")
@@ -217,7 +222,7 @@ class SessionTemplateService(
   fun deleteSessionLocationGroup(reference: String) {
     val group = getLocationGroupByReference(reference)
     if (group.sessionTemplates.isNotEmpty()) {
-      throw VSiPValidationException("Group cannot be deleted $reference because session templates are using it!")
+      throw VSiPValidationException("Location group cannot be deleted $reference because session templates are using it!")
     }
 
     val deleted = sessionLocationGroupRepository.deleteByReference(reference)
@@ -233,10 +238,6 @@ class SessionTemplateService(
     return sessionCategoryGroupRepository.findByReference(reference) ?: throw ItemNotFoundException("SessionPrisonerCategory reference:$reference not found")
   }
 
-  private fun getPrisonerIncentiveLevelGroupByReference(reference: String): SessionIncentiveLevelGroup {
-    return sessionIncentiveLevelGroupRepository.findByReference(reference) ?: throw ItemNotFoundException("SessionPrisonerIncentiveLevel reference:$reference not found")
-  }
-
   private fun getLocationGroupByReference(reference: String): SessionLocationGroup {
     return sessionLocationGroupRepository.findByReference(reference)
       ?: throw ItemNotFoundException("SessionLocationGroup reference:$reference not found")
@@ -245,6 +246,153 @@ class SessionTemplateService(
   private fun getSessionTemplate(reference: String): SessionTemplate {
     return sessionTemplateRepository.findByReference(reference)
       ?: throw TemplateNotFoundException("Template reference:$reference not found")
+  }
+
+  fun createSessionCategoryGroup(createCategorySessionGroup: CreateCategoryGroupDto): SessionCategoryGroupDto {
+    val prison = prisonConfigService.findPrisonByCode(createCategorySessionGroup.prisonCode)
+    val groupToCreate = SessionCategoryGroup(
+      prison = prison,
+      prisonId = prison.id,
+      name = createCategorySessionGroup.name,
+    )
+
+    val sessionPrisonerCategorys = createCategorySessionGroup.categories.toSet().map {
+      SessionPrisonerCategory(
+        sessionCategoryGroupId = groupToCreate.id,
+        sessionCategoryGroup = groupToCreate,
+        prisonerCategoryType = it,
+      )
+    }
+
+    groupToCreate.sessionCategories.addAll(sessionPrisonerCategorys)
+
+    val createdGroup = sessionCategoryGroupRepository.saveAndFlush(groupToCreate)
+    return SessionCategoryGroupDto(createdGroup)
+  }
+
+  fun updateSessionCategoryGroup(
+    reference: String,
+    updateCategorySessionGroup: UpdateCategoryGroupDto,
+  ): SessionCategoryGroupDto {
+    val groupToUpdate = getSessionCategoryGroupEntityByReference(reference)
+    groupToUpdate.name = updateCategorySessionGroup.name
+    groupToUpdate.sessionCategories.clear()
+
+    val sessionPrisonerCategorys = updateCategorySessionGroup.categories.toSet().map {
+      SessionPrisonerCategory(
+        sessionCategoryGroupId = groupToUpdate.id,
+        sessionCategoryGroup = groupToUpdate,
+        prisonerCategoryType = it,
+      )
+    }
+
+    groupToUpdate.sessionCategories.addAll(sessionPrisonerCategorys)
+
+    val updatedGroup = sessionCategoryGroupRepository.saveAndFlush(groupToUpdate)
+    return SessionCategoryGroupDto(updatedGroup)
+  }
+
+  fun deleteSessionCategoryGroup(reference: String) {
+    val group = getSessionCategoryGroupEntityByReference(reference)
+    if (group.sessionTemplates.isNotEmpty()) {
+      throw VSiPValidationException("Category group cannot be deleted $reference because session templates are using it!")
+    }
+
+    val deleted = sessionCategoryGroupRepository.deleteByReference(reference)
+    if (deleted == 0) {
+      throw ItemNotFoundException("Session category group not found : $reference")
+    }
+    if (deleted > 1) {
+      throw java.lang.IllegalStateException("More than one Session category group $reference was deleted!")
+    }
+  }
+
+  @Transactional(readOnly = true)
+  fun getSessionCategoryGroup(reference: String): SessionCategoryGroupDto {
+    return SessionCategoryGroupDto(getSessionCategoryGroupEntityByReference(reference))
+  }
+
+  private fun getSessionCategoryGroupEntityByReference(reference: String): SessionCategoryGroup {
+    return sessionCategoryGroupRepository.findByReference(reference)
+      ?: throw ItemNotFoundException("SessionCategoryGroupDto reference:$reference not found")
+  }
+
+  @Transactional(readOnly = true)
+  fun getSessionCategoryGroups(prisonCode: String): List<SessionCategoryGroupDto> {
+    return sessionCategoryGroupRepository.findByPrisonCode(prisonCode).map { SessionCategoryGroupDto(it) }
+  }
+
+  @Transactional(readOnly = true)
+  fun getSessionIncentiveGroups(prisonCode: String): List<SessionIncentiveLevelGroupDto> {
+    return sessionIncentiveLevelGroupRepository.findByPrisonCode(prisonCode).map { SessionIncentiveLevelGroupDto(it) }
+  }
+
+  @Transactional(readOnly = true)
+  fun getSessionIncentiveGroup(reference: String): SessionIncentiveLevelGroupDto {
+    return SessionIncentiveLevelGroupDto(getIncentiveLevelGroupByReference(reference))
+  }
+
+  private fun getIncentiveLevelGroupByReference(reference: String): SessionIncentiveLevelGroup {
+    return sessionIncentiveLevelGroupRepository.findByReference(reference) ?: throw ItemNotFoundException("SessionPrisonerIncentiveLevel reference:$reference not found")
+  }
+
+  fun createSessionIncentiveGroup(createIncentiveSessionGroup: CreateIncentiveGroupDto): SessionIncentiveLevelGroupDto {
+    val prison = prisonConfigService.findPrisonByCode(createIncentiveSessionGroup.prisonCode)
+    val groupToCreate = SessionIncentiveLevelGroup(
+      prison = prison,
+      prisonId = prison.id,
+      name = createIncentiveSessionGroup.name,
+    )
+
+    val sessionIncentiveLevels = createIncentiveSessionGroup.incentiveLevels.toSet().map {
+      SessionPrisonerIncentiveLevel(
+        sessionIncentiveGroupId = groupToCreate.id,
+        sessionIncentiveLevelGroup = groupToCreate,
+        prisonerIncentiveLevel = it,
+      )
+    }
+
+    groupToCreate.sessionIncentiveLevels.addAll(sessionIncentiveLevels)
+
+    val createdGroup = sessionIncentiveLevelGroupRepository.saveAndFlush(groupToCreate)
+    return SessionIncentiveLevelGroupDto(createdGroup)
+  }
+
+  fun updateSessionIncentiveGroup(
+    reference: String,
+    updateIncentiveSessionGroup: UpdateIncentiveGroupDto,
+  ): SessionIncentiveLevelGroupDto {
+    val groupToUpdate = getIncentiveLevelGroupByReference(reference)
+    groupToUpdate.name = updateIncentiveSessionGroup.name
+    groupToUpdate.sessionIncentiveLevels.clear()
+
+    val sessionIncentiveLevels = updateIncentiveSessionGroup.incentiveLevels.toSet().map {
+      SessionPrisonerIncentiveLevel(
+        sessionIncentiveGroupId = groupToUpdate.id,
+        sessionIncentiveLevelGroup = groupToUpdate,
+        prisonerIncentiveLevel = it,
+      )
+    }
+
+    groupToUpdate.sessionIncentiveLevels.addAll(sessionIncentiveLevels)
+
+    val updatedGroup = sessionIncentiveLevelGroupRepository.saveAndFlush(groupToUpdate)
+    return SessionIncentiveLevelGroupDto(updatedGroup)
+  }
+
+  fun deleteSessionIncentiveGroup(reference: String) {
+    val group = getIncentiveLevelGroupByReference(reference)
+    if (group.sessionTemplates.isNotEmpty()) {
+      throw VSiPValidationException("Incentive group cannot be deleted $reference because session templates are using it!")
+    }
+
+    val deleted = sessionIncentiveLevelGroupRepository.deleteByReference(reference)
+    if (deleted == 0) {
+      throw ItemNotFoundException("Incentive group  group not found : $reference")
+    }
+    if (deleted > 1) {
+      throw java.lang.IllegalStateException("More than one Incentive group $reference was deleted!")
+    }
   }
 }
